requirements.txt
fastapi
uvicorn
fastapi_users_db_sqlalchemy
fastapi_users
sqlalchemy
aiomysql 
python-dotenv